+++ example/lib/shared/const/app.dart
@@ -40,7 +40,7 @@ class App {
   // build numbers.
   static const String versionMajor = '7';
   static const String versionMinor = '3';
-  static const String versionPatch = '0';
+  static const String versionPatch = '1';
   static const String versionBuild = '01';
   static const String version = '$versionMajor.$versionMinor.$versionPatch '
       '\nBuild-$versionBuild';